feat(global-options): allow options to be shared between tasks
Also included are changes that allow multiple makefiles to be
imported creating a hierarchical structure. This is very useful for
larger projects and/or monorepos. More details can be found
in the example.

BREAKING CHANGE: fairly significant changes although none are
breaking I believe, but without a test suite it's hard to say. If this is v2,
then v3 will include a full test suite, I promise :)